requirements.txt
python-telegram-bot>=21.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
yandexcloud>=0.300.0
python-dotenv>=1.0.0